test_runner/pyproject.toml
[project]
name = "test-runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client>=5.1.1",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
